Refactor fjernet single-use interfaces fra en god del entiteter
