--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/domene/iay/AktivitetsAvtaleEntitet.java
@@ -1,245 +1,0 @@
-package no.nav.foreldrepenger.abakus.domene.iay;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.util.Objects;
-
-import javax.persistence.AttributeOverride;
-import javax.persistence.AttributeOverrides;
-import javax.persistence.Column;
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-import javax.persistence.Version;
-
-import no.nav.foreldrepenger.abakus.felles.diff.ChangeTracked;
-import no.nav.foreldrepenger.abakus.felles.diff.DiffIgnore;
-import no.nav.foreldrepenger.abakus.felles.diff.IndexKey;
-import no.nav.foreldrepenger.abakus.felles.jpa.BaseEntitet;
-import no.nav.foreldrepenger.abakus.typer.AntallTimer;
-import no.nav.foreldrepenger.abakus.typer.Stillingsprosent;
-import no.nav.vedtak.felles.jpa.tid.DatoIntervallEntitet;
-import no.nav.vedtak.konfig.Tid;
-
-@Table(name = "IAY_AKTIVITETS_AVTALE")
-@Entity(name = "AktivitetsAvtale")
-public class AktivitetsAvtaleEntitet extends BaseEntitet implements AktivitetsAvtale, IndexKey {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_AKTIVITETS_AVTALE")
-    private Long id;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(name = "yrkesaktivitet_id", nullable = false, updatable = false, unique = true)
-    private YrkesaktivitetEntitet yrkesaktivitet;
-
-    /**
-     * TODO (FC): Se om vi kan bli kvitt antallTimer. Brukes bare til å sjekke om det finnes verdi i {@link #erAnsettelsesPeriode()}.
-     */
-    @DiffIgnore
-    @Embedded
-    @AttributeOverrides(@AttributeOverride(name = "verdi", column = @Column(name = "antall_timer")))
-    private AntallTimer antallTimer;
-
-    /**
-     * TODO (FC): Se om vi kan bli kvitt antallTimerFulltid. Brukes bare til å sjekke om det finnes verdi i {@link #erAnsettelsesPeriode()}.
-     */
-    @DiffIgnore
-    @Embedded
-    @AttributeOverrides(@AttributeOverride(name = "verdi", column = @Column(name = "antall_timer_fulltid")))
-    private AntallTimer antallTimerFulltid;
-
-    @ChangeTracked
-    @Embedded
-    @AttributeOverrides(@AttributeOverride(name = "verdi", column = @Column(name = "prosentsats")))
-    private Stillingsprosent prosentsats;
-
-    @Column(name = "beskrivelse")
-    private String beskrivelse;
-
-    @Embedded
-    @ChangeTracked
-    private DatoIntervallEntitet periode;
-
-    @ChangeTracked
-    @Column(name = "siste_loennsendringsdato")
-    private LocalDate sisteLønnsendringsdato;
-
-    @Version
-    @Column(name = "versjon", nullable = false)
-    private long versjon;
-
-    /**
-     * Setter en periode brukt til overstyring av angitt periode (avledet fra saksbehandlers vurderinger). Benyttes kun transient (ved filtrering av modellen)
-     */
-    @Transient
-    private DatoIntervallEntitet overstyrtPeriode;
-
-    AktivitetsAvtaleEntitet() {
-        // hibernate
-    }
-
-    /**
-     * Deep copy ctor
-     */
-    AktivitetsAvtaleEntitet(AktivitetsAvtale aktivitetsAvtale) {
-        this.prosentsats = aktivitetsAvtale.getProsentsats();
-        this.beskrivelse = aktivitetsAvtale.getBeskrivelse();
-        this.sisteLønnsendringsdato = aktivitetsAvtale.getSisteLønnsendringsdato();
-        this.periode = aktivitetsAvtale.getPeriodeUtenOverstyring();
-    }
-
-    public AktivitetsAvtaleEntitet(AktivitetsAvtale avtale, DatoIntervallEntitet overstyrtPeriode) {
-        this(avtale);
-        this.overstyrtPeriode = overstyrtPeriode;
-    }
-
-    @Override
-    public String getIndexKey() {
-        return IndexKey.createKey(periode, sisteLønnsendringsdato);
-    }
-
-    @Override
-    public Stillingsprosent getProsentsats() {
-        return prosentsats;
-    }
-
-    void setProsentsats(Stillingsprosent prosentsats) {
-        this.prosentsats = prosentsats;
-    }
-
-    @Override
-    public BigDecimal getProsentsatsVerdi() {
-        return prosentsats == null ? null : prosentsats.getVerdi();
-    }
-
-    @Override
-    public AntallTimer getAntallTimer() {
-        return antallTimer;
-    }
-
-    void setAntallTimer(AntallTimer antallTimer) {
-        this.antallTimer = antallTimer;
-    }
-
-    @Override
-    public AntallTimer getAntallTimerFulltid() {
-        return antallTimerFulltid;
-    }
-
-    void setAntallTimerFulltid(AntallTimer antallTimerFulltid) {
-        this.antallTimerFulltid = antallTimerFulltid;
-    }
-
-    @Override
-    public DatoIntervallEntitet getPeriode() {
-        return erOverstyrtPeriode() ? overstyrtPeriode : periode;
-    }
-
-    void setPeriode(DatoIntervallEntitet periode) {
-        this.periode = periode;
-    }
-
-    /**
-     * Henter kun den originale perioden, ikke den overstyrte perioden.
-     * Bruk heller {@link #getPeriode} i de fleste tilfeller
-     *
-     * @return Hele den originale perioden, uten overstyringer.
-     */
-    @Override
-    public DatoIntervallEntitet getPeriodeUtenOverstyring() {
-        return periode;
-    }
-
-    /**
-     * @deprecated FIXME - bør fjerne intern filtrering basert på initialisert transient overstyrt periode.  Legg heller til egen Decorator klasse som filtrerer output fra entitet
-     */
-    @Deprecated
-    @Override
-    public boolean erOverstyrtPeriode() {
-        return overstyrtPeriode != null;
-    }
-
-    @Override
-    public LocalDate getSisteLønnsendringsdato() {
-        return sisteLønnsendringsdato;
-    }
-
-    @Override
-    public boolean matcherPeriode(DatoIntervallEntitet aktivitetsAvtale) {
-        return getPeriode().equals(aktivitetsAvtale);
-    }
-
-    @Override
-    public boolean getErLøpende() {
-        return Tid.TIDENES_ENDE.equals(getPeriode().getTomDato());
-    }
-
-    @Override
-    public String getBeskrivelse() {
-        return beskrivelse;
-    }
-
-    void setBeskrivelse(String beskrivelse) {
-        this.beskrivelse = beskrivelse;
-    }
-
-    void setYrkesaktivitet(YrkesaktivitetEntitet yrkesaktivitet) {
-        this.yrkesaktivitet = yrkesaktivitet;
-    }
-
-    void sisteLønnsendringsdato(LocalDate sisteLønnsendringsdato) {
-        this.sisteLønnsendringsdato = sisteLønnsendringsdato;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || !(o instanceof AktivitetsAvtaleEntitet)) return false;
-        AktivitetsAvtaleEntitet that = (AktivitetsAvtaleEntitet) o;
-        return Objects.equals(antallTimer, that.antallTimer) &&
-            Objects.equals(antallTimerFulltid, that.antallTimerFulltid) &&
-            Objects.equals(beskrivelse, that.beskrivelse) &&
-            Objects.equals(prosentsats, that.prosentsats) &&
-            Objects.equals(periode, that.periode) &&
-            Objects.equals(overstyrtPeriode, that.overstyrtPeriode) &&
-            Objects.equals(sisteLønnsendringsdato, that.sisteLønnsendringsdato);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(antallTimer, antallTimerFulltid, beskrivelse, prosentsats, periode, overstyrtPeriode, sisteLønnsendringsdato);
-    }
-
-    @Override
-    public String toString() {
-        return getClass().getSimpleName() + "<" + //$NON-NLS-1$
-            "antallTimer=" + antallTimer + //$NON-NLS-1$
-            ", antallTimerFulltid=" + antallTimerFulltid + //$NON-NLS-1$
-            ", periode=" + periode + //$NON-NLS-1$
-            ", overstyrtPeriode=" + overstyrtPeriode + //$NON-NLS-1$
-            ", prosentsats=" + prosentsats + //$NON-NLS-1$
-            ", beskrivelse=" + beskrivelse + //$NON-NLS-1$
-            ", sisteLønnsendringsdato=" + sisteLønnsendringsdato + //$NON-NLS-1$
-            '>';
-    }
-
-    boolean hasValues() {
-        return prosentsats != null || periode != null;
-    }
-
-    @Override
-    public boolean erAnsettelsesPeriode() {
-        return (antallTimer == null || antallTimer.getVerdi() == null)
-            && (antallTimerFulltid == null || antallTimerFulltid.getVerdi() == null)
-            && (prosentsats == null || prosentsats.erNulltall())
-            && sisteLønnsendringsdato == null;
-    }
-
-}
