+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/domene/iay/arbeidsforhold/ArbeidsforholdInformasjonBuilder.java
@@ -15,20 +15,20 @@ import no.nav.vedtak.util.Tuple;
 
 public class ArbeidsforholdInformasjonBuilder {
 
-    private final ArbeidsforholdInformasjonEntitet kladd;
+    private final ArbeidsforholdInformasjon kladd;
     private final List<Tuple<Arbeidsgiver, Tuple<InternArbeidsforholdRef, InternArbeidsforholdRef>>> erstattArbeidsforhold = new ArrayList<>();
     private final List<Tuple<Arbeidsgiver, Tuple<InternArbeidsforholdRef, InternArbeidsforholdRef>>> reverserteErstattninger = new ArrayList<>();
 
-    private ArbeidsforholdInformasjonBuilder(ArbeidsforholdInformasjonEntitet kladd) {
+    private ArbeidsforholdInformasjonBuilder(ArbeidsforholdInformasjon kladd) {
         this.kladd = kladd;
     }
 
     public static ArbeidsforholdInformasjonBuilder oppdatere(ArbeidsforholdInformasjon oppdatere) {
-        return new ArbeidsforholdInformasjonBuilder(new ArbeidsforholdInformasjonEntitet(oppdatere));
+        return new ArbeidsforholdInformasjonBuilder(new ArbeidsforholdInformasjon(oppdatere));
     }
 
     public static ArbeidsforholdInformasjonBuilder builder(Optional<ArbeidsforholdInformasjon> arbeidsforholdInformasjon) {
-        var arbeidInfo = arbeidsforholdInformasjon.map(ai -> new ArbeidsforholdInformasjonEntitet(ai)).orElseGet(() -> new ArbeidsforholdInformasjonEntitet());
+        var arbeidInfo = arbeidsforholdInformasjon.map(ai -> new ArbeidsforholdInformasjon(ai)).orElseGet(() -> new ArbeidsforholdInformasjon());
         return new ArbeidsforholdInformasjonBuilder(arbeidInfo);
     }
 
@@ -96,7 +96,7 @@ public class ArbeidsforholdInformasjonBuilder {
         return this;
     }
 
-    public ArbeidsforholdInformasjonEntitet build() {
+    public ArbeidsforholdInformasjon build() {
         return kladd;
     }
